requirements.txt
numpy
pillow
torch
torchvision
tensorflow
pytest
ultralytics
